pyproject.toml
[project]
name = "emailkasten"
version = "0.0.1"
description = "A open-source self-hosted mail archive using the django framework."
authors = [
    {name = "David Aderbauer & The Emailkasten Contributors", email = "18516197-Dacid99@users.noreply.gitlab.com"}
]
license = {text = "AGPL-3.0-or-later"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "django-extensions (>=4.1,<5.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-health-check (>=3.18.3,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-dirtyfields (>=1.9.7,<2.0.0)",
    "django-allauth (>=65.7.0,<66.0.0)",
    "django-constance (>=4.3.2,<5.0.0)",
    "dj-rest-auth (>=7.0.1,<8.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "exchangelib (>=5.5.1,<6.0.0)",
    "imap-tools (>=1.10.0,<2.0.0)",
    "mysqlclient (>=2.2.7, <3.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "django-bootstrap5 (>=25.1,<26.0)",
    "fontawesomefree (>=6.6.0,<7.0.0)",
    "django-crispy-forms (>=2.4,<3.0)",
    "crispy-bootstrap5 (>=2025.4,<2026.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "django-robots (>=6.1,<7.0)",
    "tomli (>=2.2.1,<3.0.0)",
    "gssapi (>=1.9.0,<2.0.0)",
    "krb5 (>=0.7.1,<0.8.0)",
    "vobject (>=0.9.9,<0.10.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<4.0"

[tool.poetry.group.dev]
optional=true

[tool.poetry.group.dev.dependencies]
django-stubs = "^5.1.3"
djangorestframework-stubs = "^3.15.3"
django-filter-stubs = "^0.1.3"
black = "^25.1.0"
ruff = "^0.12.1"
autotyping = "^24.9.0"
mypy = "^1.15.0"
lxml = "^6.0.0"
pylint = "^3.3.6"
pytest = "^8.3.5"
pylint-django = "^2.6.1"
pytest-django = "^4.11.0"
pytest-mock = "^3.14.0"
pytest-mypy = "^1.0.1"
pytest-cov = "^6.1.0"
pytest-html = "^4.1.1"
pyfakefs = "^5.8.0"
factory-boy = "^3.3.3"
model-bakery = "^1.20.4"
pylintfileheader = "^1.0.0"
freezegun = "^1.5.4"
djlint = "^1.36.4"
codespell = "^2.4.1"
doc8 = "^2.0.0"
celery-stubs = "^0.1.3"
wlc = "^1.15"

[tool.poetry.group.docs]
optional=true

[tool.poetry.group.docs.dependencies]
sphinx = "^8.2.3"
sphinxcontrib-napoleon = "^0.7"
sphinxcontrib-django = "^2.5"
sphinx-autodoc-typehints = "^3.1.0"
myst-parser = "^4.0.1"
pydeps = "^3.0.1"
pydot = "^4.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
